renaming signature file to be more descriptive
